docs: update getting started guide and dependencies
