+++ src/error.ts
@@ -8,20 +8,8 @@ export class AppError extends Error {
   }
 }
 
-export class UnauthorizedError extends AppError {
-  constructor(message: string) {
-    super(message, 401);
-  }
-}
-
 export class NotFoundError extends AppError {
   constructor(message: string) {
     super(message, 404);
-  }
-}
-
-export class ValidationError extends AppError {
-  constructor(message: string) {
-    super(message, 422);
   }
 }
