+++ src/routes/api/federations/federations.service.ts
@@ -1,6 +1,5 @@
 import type { DataStoreType } from "../../../data/store";
-import type { Pagination } from "../../../types";
-import { buildPagination } from "../../../utils/helpers";
+import { type Pagination, buildPagination } from "../../../utils/helpers";
 import { configuration } from "../../../configuration";
 import type { GetFederationsType, GetFederationMeetsQueryType } from "./federations.validation";
