+++ src/routes/api/rankings/rankings.service.ts
@@ -1,7 +1,6 @@
 import type { DataStoreType } from "../../../data/store";
 import type { AppData, Entry, RankMetric } from "../../../data/types";
-import type { Pagination } from "../../../types";
-import { type Units, buildPagination, inUnits } from "../../../utils/helpers";
+import { type Pagination, type Units, buildPagination, inUnits } from "../../../utils/helpers";
 import { configuration } from "../../../configuration";
 import type {
   GetFilteredRankingsParamType,
